+++ src/pages/history.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { useNavigate, useParams } from "react-router";
 import {
+  useId,
   useMenu,
   useProfile,
   useSearch,
@@ -21,17 +22,18 @@ export default function History() {
   const [history, setHistory] = useHistory();
   const [width] = useWidth();
   const [selected, setSelected] = useSelected();
-  ![0, 1].includes(selected) &&
-    setSelected(0);
+  const [id, setId] = useId();
   if (!(width < 760)) {
     navigate(`/profile/${params.id}`);
   } else {
     !menu && setMenu(true);
     history !== window.location.pathname &&
       setHistory(window.location.pathname);
-    (profile !== Number(params.id) || "self") && 
+    (profile !== Number(params.id) || "self") &&
       setProfile(Number(params.id) || "self");
     search && setSearch(false);
+    selected && setSelected(0);
+    id && setId(0);
   }
   return <div />;
 }