+++ src/pages/search.tsx
@@ -3,6 +3,7 @@ import { Box, Paper } from "@mui/material";
 import Empty from "../components/empty";
 import {
   useData,
+  useId,
   useMenu,
   useSearch,
   useSelected,
@@ -17,10 +18,12 @@ export default function Search() {
   const [width] = useWidth();
   const [selected, setSelected] = useSelected();
   const [, setTitle] = useTitle();
+  const [id, setId] = useId();
   document.title = "Search | Metahkg";
   history !== window.location.pathname &&
     setHistory(window.location.pathname + window.location.search);
   !menu && setMenu(true);
+  id && setId(0);
   if (!search) {
     setSearch(true);
     data && setData([]);